grouping support and fixed bugs when adding new dev weeks
The grouping support uses a method in master that isn't in the sproutcore gem yet.
Then it supplies more properties to list view so the list view knows which items are group headers.
Currently the list view doesn't display the group headers any differently.

The new dev week bugs were in the matrix view.  It didn't correclty handle null properties.